pyproject.toml
[tool.poetry]
name = "iroha-kaede"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.1.5"
chromedriver-binary = "101.0.4951.41.0"
python-dotenv = "^0.20.0"
autopep8 = "^1.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
